LayoutTests/imported/w3c/web-platform-tests/scroll-animations/view-timelines/view-timeline-range-large-subject-expected.txt


FAIL View timeline with range set via delays. assert_equals: Effect at the midpoint of the active range: cover 0% to cover 100% expected "0.5" but got "0.493333"
